Fix CTRL+P in editor
This fixes CTRL+P in the editor to "play" the current script. I
tested it for the default Ace keybindings and for the Vim ones.

The significant change is twofold. First, the relevant props for
the active buffer and for the scriptRun function are passed as
props from the EditActivity to its child Editor. The former prop
had to be aliased to activityActiveBuffer due to conflicts
surfaced in testing. This is then wrapped in a getActiveBuffer()
helper in render(). The next change is to add a "commands" prop
which is an array of commands associated with keys. The sole
member of this array is a mapping of CTRL+P to a call to
scriptRun with the active buffer.

Next, for Vim, the <C-p> mapping found in the defaults is just
removed. The Vim keybindings take priority in Ace, and removing
that binding from Vim forces it to fall back to the binding list,
which includes the aforementioned CTRL+P mapping.