Add react-router (fix #85, fix #70)
This adds react-router and supporting packages and integrates them
into the root reducer and middleware. It also introduces a URL
handler bound to state, which sets the URL according to the path
of the last script loaded into the REPL, and attempts to load the
script in the REPL on page load if the URL contains such a path.
An async action is also added to recursively open the directory
hierarchy containing the script. Some support utilities are added
to convert between a file path and a URL path. Finally, the fetch
for REPL endpoints now looks in both / and /maiden before failing.