Refactor expand behavior
Expand behavior caused "flashing" when loading file in another
directory. This is refactored to check whether a path is loaded
into the URL handler (which should always be the case except on
initial page load) and whether the path matches the one already
loaded. inSameDir is factored out, no longer used, and removed.